+++ feishukit/__init__.py
@@ -1,7 +1,27 @@
-from .feishu_api import FeishuRuntimeError, FeishuAPI
+from .feishu_api import FeishuAPI, FeishuRuntimeError
 from .feishu_bitable import Bitable
 from .feishu_doc import FeishuDoc
 from .feishu_driver import FeishuDriver
+from .feishu_spreadsheet import FeishuSpreadsheet
+from .feishu_user import (
+    FeishuAuthError,
+    FeishuAuthorizationRejectedError,
+    FeishuDeviceAuthTimeoutError,
+    FeishuUser,
+    FeishuUserDeviceAuth,
+)
 
 __version__ = "0.0.4"
-__all__ = ["FeishuAPI", "FeishuRuntimeError", "Bitable", "FeishuDoc", "FeishuDriver"]
+__all__ = [
+    "FeishuAPI",
+    "FeishuUser",
+    "FeishuUserDeviceAuth",
+    "FeishuRuntimeError",
+    "FeishuAuthError",
+    "FeishuDeviceAuthTimeoutError",
+    "FeishuAuthorizationRejectedError",
+    "Bitable",
+    "FeishuDoc",
+    "FeishuDriver",
+    "FeishuSpreadsheet",
+]